+++ server/requestorserver/server.go
@@ -101,8 +101,8 @@ func (s *Server) startServer(handler http.Handler, name, addr string, port int, 
 		Handler:   handler,
 		TLSConfig: tlsConf,
 		// See https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/
-		ReadTimeout:  server.ReadTimeout,
-		WriteTimeout: server.WriteTimeout,
+		// Write timeouts are handled per request using middleware (to exclude SSE endpoints)
+		ReadTimeout: server.ReadTimeout,
 	}
 
 	go func() {
@@ -193,6 +193,7 @@ func (s *Server) Handler() http.Handler {
 
 	router.Group(func(r chi.Router) {
 		r.Use(server.SizeLimitMiddleware)
+		r.Use(server.TimeoutMiddleware([]string{"/statusevents"}, server.WriteTimeout))
 		r.Use(cors.New(corsOptions).Handler)
 		if s.conf.Verbose >= 2 {
 			r.Use(server.LogMiddleware("requestor", log))
@@ -217,6 +218,7 @@ func (s *Server) Handler() http.Handler {
 
 	router.Group(func(r chi.Router) {
 		r.Use(server.SizeLimitMiddleware)
+		r.Use(server.TimeoutMiddleware(nil, server.WriteTimeout))
 		r.Use(cors.New(corsOptions).Handler)
 		if s.conf.Verbose >= 2 {
 			r.Use(server.LogMiddleware("revocation", log))